Refactor progress to Java 17

+++ common/src/main/java/org/apache/commons/math3/exception/MultiDimensionMismatchException.java
@@ -79,13 +79,13 @@ public class MultiDimensionMismatchException extends MathIllegalArgumentExceptio
      * @return the wrong dimension stored at {@code index}.
      */
     public int getWrongDimension(int index) {
-        return wrong[index].intValue();
+        return wrong[index];
     }
     /**
      * @param index Dimension index.
      * @return the expected dimension stored at {@code index}.
      */
     public int getExpectedDimension(int index) {
-        return expected[index].intValue();
+        return expected[index];
     }
 }
