+++ desktop/src/main/java/org/geogebra/desktop/gui/dialog/DecorationListRenderer.java
@@ -30,7 +30,7 @@ public class DecorationListRenderer extends JPanel implements ListCellRenderer {
 			int index, boolean isSelected, boolean cellHasFocus) {
 		// Get the selected index. (The index param isn't
 		// always valid, so just use the value.)
-		int selectedIndex = ((Integer) value).intValue();
+		int selectedIndex = (Integer) value;
 		this.id = selectedIndex;
 		if (isSelected) {
 			setBackground(ThemeD.color(ColorKeys.OUTLINE_LIGHT));
