+++ desktop/src/main/java/org/geogebra/desktop/gui/toolbar/ModeCellRenderer.java
@@ -59,7 +59,7 @@ public class ModeCellRenderer extends DefaultTreeCellRenderer
 			Object ob = node.getUserObject();
 			if (ob instanceof Integer) {
 				// mode node
-				int mode = ((Integer) ob).intValue();
+				int mode = (Integer) ob;
 				handleModeNode(mode);
 			} else {
 				// root node
@@ -74,7 +74,7 @@ public class ModeCellRenderer extends DefaultTreeCellRenderer
 				DefaultMutableTreeNode childNode = (DefaultMutableTreeNode) node
 						.getFirstChild();
 				Object ob = childNode.getUserObject();
-				handleModeNode(((Integer) ob).intValue());
+				handleModeNode((Integer) ob);
 			}
 		}
 
@@ -86,7 +86,7 @@ public class ModeCellRenderer extends DefaultTreeCellRenderer
 	@Override
 	public Component getListCellRendererComponent(JList list, Integer value,
 			int index, boolean isSelected, boolean cellHasFocus) {
-		handleModeNode(value.intValue());
+		handleModeNode(value);
 		handleSelection(isSelected);
 		return this;
 	}
